pyproject.toml
[project]
name = "pinjected-anthropic"
version = "0.1.11"
description = "Add your description here"
authors = [
    { name = "proboscis", email = "nameissoap@gmail.com" }
]
dependencies = [
    "pinjected>=0.2.189",
    "anthropic>=0.34.1",
    "pillow>=10.4.0",
    "pandas>=2.2.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pinjected_anthropic"]
